Move `createRandomChatIn()` and `sendRandomMessageTo()` to `MessageTestEnv`.
